requirements.txt
Django>=4.2,<5.1
wagtail>=6.1,<6.2
psycopg2
pandas
certifi
social-auth-app-django
requests-oauthlib
django-debug-toolbar
djangorestframework-simplejwt
django-cors-headers
google-api-python-client
wagtail_modeladmin
django-autocomplete-light
djangorestframework